+++ packages/analytics/src/index.ts
@@ -20,3 +20,4 @@ export * from './relativeValue';
 export * from './capm';
 export * from './insider';
 export * from './eightK';
+export * from './scoring';